ECO-008 Move model package closer to the services
