+++ src/main/java/com/ecore/roles/repository/MembershipRepository.java
@@ -1,6 +1,6 @@
 package com.ecore.roles.repository;
 
-import com.ecore.roles.model.Membership;
+import com.ecore.roles.service.model.Membership;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
