+++ src/main/java/com/ecore/roles/service/RolesService.java
@@ -1,6 +1,6 @@
 package com.ecore.roles.service;
 
-import com.ecore.roles.model.Role;
+import com.ecore.roles.service.model.Role;
 
 import java.util.List;
 import java.util.UUID;
