+++ src/main/java/com/ecore/roles/service/impl/MembershipsServiceImpl.java
@@ -5,8 +5,8 @@ import com.ecore.roles.client.model.User;
 import com.ecore.roles.exception.InvalidArgumentException;
 import com.ecore.roles.exception.ResourceExistsException;
 import com.ecore.roles.exception.ResourceNotFoundException;
-import com.ecore.roles.model.Membership;
-import com.ecore.roles.model.Role;
+import com.ecore.roles.service.model.Membership;
+import com.ecore.roles.service.model.Role;
 import com.ecore.roles.repository.MembershipRepository;
 import com.ecore.roles.service.MembershipsService;
 import com.ecore.roles.service.RolesService;
