+++ src/main/java/com/ecore/roles/web/dto/MembershipDto.java
@@ -1,7 +1,7 @@
 package com.ecore.roles.web.dto;
 
-import com.ecore.roles.model.Membership;
-import com.ecore.roles.model.Role;
+import com.ecore.roles.service.model.Membership;
+import com.ecore.roles.service.model.Role;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
