+++ src/main/java/com/ecore/roles/web/dto/RoleDto.java
@@ -1,6 +1,6 @@
 package com.ecore.roles.web.dto;
 
-import com.ecore.roles.model.Role;
+import com.ecore.roles.service.model.Role;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
