+++ src/test/java/com/ecore/roles/api/MembershipsApiTests.java
@@ -1,7 +1,7 @@
 package com.ecore.roles.api;
 
-import com.ecore.roles.model.Membership;
-import com.ecore.roles.model.Role;
+import com.ecore.roles.service.model.Membership;
+import com.ecore.roles.service.model.Role;
 import com.ecore.roles.repository.MembershipRepository;
 import com.ecore.roles.utils.RestAssuredHelper;
 import com.ecore.roles.web.dto.MembershipDto;
